Update QRCodeCard.ts
Do we think the QRcode color scheme should match the color scheme of Solid? Tested the qrcode with my phone and the change of the colors does not seem to change any information about the URI. Snapchat and Instagram also have matching color schemes to their QR cards as well.