Update QRCodeCard.ts
Tested on Android 12 and Android 14, this is a proof of concept with minimal strings. Tim advised me about the complex nature of qrcodes and information storing. I left the const URL:string alone. Many of you may attempt to rename that to URI and even the string noted as 'URL:', but I simply did not touch that because I'm not sure if the vCard will accept and if you feel URI is a better term for solid vocab/naming convention consistency, be my guest.