Update wrapped-profileFormText.ts (#62)
I believe that since it says "Short name for chats, etc" a reasonable starting place may be a 20 character limit. If characters are increased, this will lead to spammed/ugly profile homepages. I tested the 20 character limit with the default browser setting at 100% zoom function. It does not interfere with the cards aesthetic. It seems like it stays in its container and looks encapsulated. Also checked this out on the smartphone and it seems safe. I also noticed that there are two wrapped-profileFormText.ts files that seem identical. Place this where you will. If this is not fixed, users will exploit this and spam to infinity.